requirements.txt
python-slugify
youtube-dl

